Fix couple of powerup values

+++ Gameplay/Levelling/PowerUps/Player/ExperienceUp.cs
@@ -1,4 +1,4 @@
 ﻿namespace Gameplay.Levelling.PowerUps.Player;
 
 /// <param name="Value">Proportional increase. Accumulated additively.</param>
-public record ExperienceUp(int Value) : IPlayerPowerUp;
+public record ExperienceUp(float Value) : IPlayerPowerUp;
